backend/requirements.txt
certifi==2021.5.30
importlib-metadata==4.6.4
packaging==21.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
python-dotenv==0.19.1
toml==0.10.2
urllib3==1.26.6
zipp==3.5.0
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
djoser==2.0.1
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==21.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
pillow==8.3.2
sorl-thumbnail==12.6.3